+++ analyzers/haskell/analyzer/cabal.go
@@ -1,71 +1,26 @@
 package analyzer
 
 import (
-	"errors"
-	"github.com/fossas/fossa-cli/exec"
-	"github.com/fossas/fossa-cli/files"
+	"github.com/fossas/fossa-cli/buildtools/cabal"
 	"github.com/fossas/fossa-cli/graph"
 	"github.com/fossas/fossa-cli/module"
 	"github.com/fossas/fossa-cli/pkg"
-	"github.com/mitchellh/mapstructure"
-	"path/filepath"
 )
 
-const CabalPlanRelPath = "dist-newstyle/cache/plan.json"
-
-type CabalPlan struct {
-	InstallPlans []InstallPlan `mapstructure:"install-plan"`
-}
-
-type InstallPlan struct {
-	Type    string   `mapstructure:"type"`
-	Id      string   `mapstructure:"id"`
-	Name    string   `mapstructure:"pkg-name"`
-	Version string   `mapstructure:"pkg-version"`
-
-	Components map[string]Component `mapstructure:"components"` // Not always present
-	Depends    []string             `mapstructure:"depends"`    // Dependencies can be defined here or in Components.*.Depends
-	Style      string               `mapstructure:"style"`      // Only exists for packages with type `configured`
-}
-
-type Component struct {
-	Depends []string `mapstructure:"depends"`
-}
-
 func AnalyzeCabal(m module.Module) (graph.Deps, error) {
-	cabalPlanPath := filepath.Join(m.Dir, CabalPlanRelPath)
+	plan, err := cabal.GetSolverPlan(m.Dir)
 
-	// If plan.json doesn't exist, generate it
-	if exists, _ := files.Exists(cabalPlanPath); !exists {
-		_, _, err := exec.Run(exec.Cmd{
-			Name: "cabal",
-			Argv: []string{"new-build", "--dry-run"},
-		})
-
-		if err != nil {
-			return graph.Deps{}, err
-		}
-	}
-
-	if exists, _ := files.Exists(cabalPlanPath); !exists {
-		// TODO: fallback to another strategy?
-		return graph.Deps{}, errors.New("couldn't find or generate cabal solver plan")
-	}
-
-	// Parse cabal new-build's build plan
-	var rawPlan map[string]interface{}
-	var plan    CabalPlan
-
-	if err := files.ReadJSON(&rawPlan, cabalPlanPath); err != nil {
-		return graph.Deps{}, err
-	}
-	if err := mapstructure.Decode(rawPlan, &plan); err != nil {
+	if err != nil {
 		return graph.Deps{}, err
 	}
 
+	return AnalyzeCabalPure(plan), nil
+}
+
+func AnalyzeCabalPure(plan cabal.Plan) graph.Deps {
 	// Key a map of install plans by their IDs
 	// this is used when building out the dependency graph
-	var installPlans = make(map[string]InstallPlan)
+	var installPlans = make(map[string]cabal.InstallPlan)
 	for _, p := range plan.InstallPlans {
 		installPlans[p.Id] = p
 	}
@@ -90,25 +45,17 @@ func AnalyzeCabal(m module.Module) (graph.Deps, error) {
 		Transitive: dependencyGraph,
 	}
 
-	return deps, nil
+	return deps
 }
 
-func InstallPlanToID(plan InstallPlan) pkg.ID {
-	return pkg.ID{
-		Type:     pkg.Haskell,
-		Name:     plan.Name,
-		Revision: plan.Version,
-	}
-}
-
-func InstallPlanToPackage(installPlans map[string]InstallPlan, plan InstallPlan) pkg.Package {
+func InstallPlanToPackage(installPlans map[string]cabal.InstallPlan, plan cabal.InstallPlan) pkg.Package {
 	var imports []pkg.Import
 
 	for _, depId := range plan.Depends {
 		dep := installPlans[depId]
 
 		imports = append(imports, pkg.Import{
-			Resolved: InstallPlanToID(dep),
+			Resolved: cabal.InstallPlanToID(dep),
 		})
 	}
 
@@ -117,13 +64,13 @@ func InstallPlanToPackage(installPlans map[string]InstallPlan, plan InstallPlan)
 			dep := installPlans[depId]
 
 			imports = append(imports, pkg.Import{
-				Resolved: InstallPlanToID(dep),
+				Resolved: cabal.InstallPlanToID(dep),
 			})
 		}
 	}
 
 	return pkg.Package{
-		ID: InstallPlanToID(plan),
+		ID: cabal.InstallPlanToID(plan),
 		Imports: imports,
 	}
 }
